DTO and Controller

+++ src/main/java/src/studet/StudentServiceImpl.java
@@ -6,20 +6,26 @@ import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.transaction.Transactional;
 import src.universite.University;
+import src.universite.UniversityRepository;
 import src.utils.Specialite;
 
 @Service
+@Transactional
 public class StudentServiceImpl implements StudentService{
 	
 	private StudentRepository studentRepo;
+	private final UniversityRepository universityRepo;
+
+	 public StudentServiceImpl(StudentRepository studentRepo,
+             UniversityRepository universityRepo) {
+			this.studentRepo = studentRepo;
+			this.universityRepo = universityRepo;
+	 }
 	
 	
-
-	public StudentServiceImpl(StudentRepository studentRepo) {
-		this.studentRepo=studentRepo;
-	}
-
 
     @Override
     public Student createStudent(Student student) {
@@ -31,17 +37,21 @@ public class StudentServiceImpl implements StudentService{
 
     @Override
     public List<Student> getAllStudents() {
-        return studentRepo.findAll();
+    	 List<Student> students = studentRepo.findAll();    	
+    	    students.forEach(s -> s.getUniversity()); // force le chargement lazy
+    	    return students;        
     }
 
     @Override
     public Optional<Student> getStudentById(UUID id) {
-        return studentRepo.findById(id);
+    	Optional<Student> student = studentRepo.findById(id);
+        student.ifPresent(s -> s.getUniversity()); // force le chargement lazy
+        return student;        
     }
 
     @Override
     public Optional<Student> getStudentByEmail(String email) {
-        return studentRepo.findStudentByEmail(email);
+        return studentRepo.findByEmail(email);
     }
 
     @Override
@@ -56,7 +66,11 @@ public class StudentServiceImpl implements StudentService{
 
     @Override
     public void deleteStudent(UUID id) {
+    	if(!studentRepo.existsById(id)) {
+    	    throw new EntityNotFoundException("Student introuvable");
+    	}
     	studentRepo.deleteById(id);
+
     }
 
 
@@ -65,6 +79,20 @@ public class StudentServiceImpl implements StudentService{
 		studentRepo.deleteAll();
 		
 	}
+	public Student assignStudentToUniversity(UUID studentId, UUID universityId) {
+
+        Student student = studentRepo.findById(studentId)
+                .orElseThrow(() -> new RuntimeException("Student introuvable"));
+
+        University university = universityRepo.findById(universityId)
+                .orElseThrow(() -> new RuntimeException("University introuvable"));
+
+        university.addStudent(student); 
+        student.setUniversity(university);
+        student.getUniversity();
+        
+        return studentRepo.save(student);
+    }
 
 
 }
